examples/sentimark_requirements.txt
dominate
newspaper3k
textblob
